Fill majority
